Revert "Internal: Cherry-pick PR 33892 to main with conflicts Further VQA twe…"
This reverts commit ac4c5d379c272118c9c28eeac40a97f58a2c392a.
