+++ modules/editor-one/assets/js/sidebar-navigation/components/collapsed-menu/collapsed-menu-item-tooltip.js
@@ -1,18 +1,14 @@
-import { Tooltip, ListItem } from '@elementor/ui';
+import { Tooltip } from '@elementor/ui';
 import PropTypes from 'prop-types';
-import { MenuItemButton, MenuIcon } from '../shared';
+import { CollapsedIconButton } from './styled-components';
 
-const CollapsedMenuItemTooltip = ( { item, isActive, onClick, IconComponent, onMouseEnter } ) => {
+const CollapsedMenuItemTooltip = ( { item, isActive, onClick, IconComponent } ) => {
 	return (
-		<ListItem disablePadding dense disableGutters onMouseEnter={ onMouseEnter }>
-			<Tooltip title={ item.label } placement="right">
-				<MenuItemButton onClick={ onClick } selected={ isActive } sx={ { height: 36 } }>
-					<MenuIcon>
-						<IconComponent />
-					</MenuIcon>
-				</MenuItemButton>
-			</Tooltip>
-		</ListItem>
+		<Tooltip title={ item.label } placement="right">
+			<CollapsedIconButton onClick={ onClick } isHighlighted={ isActive }>
+				<IconComponent />
+			</CollapsedIconButton>
+		</Tooltip>
 	);
 };
 
@@ -21,7 +17,7 @@ CollapsedMenuItemTooltip.propTypes = {
 	isActive: PropTypes.bool.isRequired,
 	onClick: PropTypes.func.isRequired,
 	IconComponent: PropTypes.elementType.isRequired,
-	onMouseEnter: PropTypes.func.isRequired,
 };
 
 export default CollapsedMenuItemTooltip;
+
