+++ modules/editor-one/assets/js/sidebar-navigation/components/collapsed-menu/index.js
@@ -1,3 +1,4 @@
 export { default as SidebarCollapsedMenu } from './sidebar-collapsed-menu';
 export { default as SidebarCollapsedMenuItem } from './sidebar-collapsed-menu-item';
+export * from './styled-components';
