+++ modules/editor-one/assets/js/sidebar-navigation/components/collapsed-menu/sidebar-collapsed-menu-item.js
@@ -1,8 +1,9 @@
-import { useState } from '@wordpress/element';
+import { useRef } from '@wordpress/element';
 import PropTypes from 'prop-types';
 import { DEFAULT_ICON, ICON_MAP } from '../shared';
 import CollapsedMenuItemPopover from './collapsed-menu-item-popover';
 import CollapsedMenuItemTooltip from './collapsed-menu-item-tooltip';
+import { CollapsedMenuItemContainer } from './styled-components';
 
 const SidebarCollapsedMenuItem = ( {
 	item,
@@ -13,8 +14,8 @@ const SidebarCollapsedMenuItem = ( {
 	onOpenPopover,
 	onClosePopover,
 } ) => {
-	const [ anchorEl, setAnchorEl ] = useState( null );
-	const hasChildren = !! children?.length;
+	const anchorRef = useRef( null );
+	const hasChildren = children && children.length > 0;
 	const IconComponent = ICON_MAP[ item.icon ] || DEFAULT_ICON;
 
 	const handleMouseEnter = () => {
@@ -32,19 +33,17 @@ const SidebarCollapsedMenuItem = ( {
 	};
 
 	return (
-		<>
+		<CollapsedMenuItemContainer ref={ anchorRef } onMouseEnter={ handleMouseEnter }>
 			{ hasChildren ? (
 				<CollapsedMenuItemPopover
 					item={ item }
 					children={ children }
 					activeChildSlug={ activeChildSlug }
 					isPopoverOpen={ isPopoverOpen }
-					anchorEl={ anchorEl }
+					anchorEl={ anchorRef.current }
 					onClose={ onClosePopover }
 					IconComponent={ IconComponent }
 					isActive={ isActive }
-					onMouseEnter={ handleMouseEnter }
-					anchorRef={ setAnchorEl }
 				/>
 			) : (
 				<CollapsedMenuItemTooltip
@@ -52,10 +51,9 @@ const SidebarCollapsedMenuItem = ( {
 					isActive={ isActive }
 					onClick={ handleClick }
 					IconComponent={ IconComponent }
-					onMouseEnter={ handleMouseEnter }
 				/>
 			) }
-		</>
+		</CollapsedMenuItemContainer>
 	);
 };
