+++ modules/editor-one/assets/js/sidebar-navigation/components/header/styled-components.js
@@ -2,12 +2,13 @@ import { Box, IconButton, Typography, styled } from '@elementor/ui';
 
 export const HeaderContainer = styled( Box )( ( { theme } ) => ( {
 	position: 'relative',
-	marginLeft: theme.spacing( 2 ),
-	marginRight: theme.spacing( 2 ),
+	paddingLeft: theme.spacing( 2 ),
+	paddingRight: theme.spacing( 2 ),
 	height: 80,
 	borderBottom: `1px solid ${ theme.palette.divider }`,
 	display: 'flex',
 	alignItems: 'center',
+	marginBottom: theme.spacing( 1 ),
 } ) );
 
 export const HeaderContent = styled( Box )( ( { theme } ) => ( {
