+++ modules/editor-one/assets/js/sidebar-navigation/components/menu/index.js
@@ -1,3 +1,4 @@
 export { default as SidebarMenu } from './sidebar-menu';
 export { default as SidebarMenuItem } from './sidebar-menu-item';
+export * from './styled-components';
