+++ modules/element-manager/assets/js/app-editor-one/hooks/useElementManager.js
@@ -1,4 +1,4 @@
-import { useEffect, useState, useCallback, useRef } from 'react';
+import { useEffect, useState, useCallback } from 'react';
 import {
 	getAdminAppData,
 	getUsageWidgets,
@@ -7,7 +7,6 @@ import {
 } from '../../api';
 
 export const useElementManager = () => {
-	const isInitialMount = useRef( true );
 	const [ isLoading, setIsLoading ] = useState( true );
 	const [ widgets, setWidgets ] = useState( [] );
 	const [ promotionWidgets, setPromotionWidgets ] = useState( [] );
@@ -107,7 +106,6 @@ export const useElementManager = () => {
 
 			setPlugins( pluginsData );
 			setIsLoading( false );
-			setChangeProgress( ( prev ) => ( { ...prev, isUnsavedChanges: false } ) );
 		};
 
 		loadData();
@@ -117,12 +115,6 @@ export const useElementManager = () => {
 		if ( isLoading ) {
 			return;
 		}
-
-		if ( isInitialMount.current ) {
-			isInitialMount.current = false;
-			return;
-		}
-
 		setChangeProgress( ( prev ) => ( { ...prev, isUnsavedChanges: true } ) );
 	}, [ widgetsDisabled, widgetsRoleRestrictions, isLoading ] );
